init setting2
