+++ .gitignore
@@ -1,0 +1,6 @@
+# Java Spring
+.classpath
+.project
+.settings/
+target/
+bin/
