--- src/main/java/com/example/trendyTracker/TrendyTrackerApplication.java
@@ -1,13 +1,0 @@
-package com.example.trendyTracker;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TrendyTrackerApplication {
-	// http://localhost:8080/swagger-ui/index.html
-	public static void main(String[] args) {
-		SpringApplication.run(TrendyTrackerApplication.class, args);
-	}
-
-}
